backend/requirements.txt
# Framework y utilidades
flask>=3.0.0
flask-cors>=4.0.0
python-dotenv>=1.0.0
werkzeug>=3.0.0
openai>=1.0.0
pytest>=7.0.0
pytest-asyncio>=0.21.0
pytest-cov>=4.1.0
pytest-mock>=3.12.0

# Logging y utilidades
python-json-logger==2.0.7

# Seguridad
pyjwt==2.8.0
cryptography==41.0.7
flask-limiter==3.5.0 